Refactor code into the relevant files and add as a loss function
